--- src/main/kotlin/omock/model/Black.kt
@@ -1,3 +1,0 @@
-package omock.model
-
-class Black(private val stone: Stone) : Placed(stone)
