--- src/main/kotlin/omock/model/Clear.kt
@@ -1,3 +1,0 @@
-package omock.model
-
-class Clear(private val stone: Stone) : UnPlaced(stone)
