--- src/main/kotlin/omock/model/White.kt
@@ -1,3 +1,0 @@
-package omock.model
-
-class White(private val stone: Stone) : Placed(stone)
