+++ src/main/kotlin/omock/model/position/Column.kt
@@ -1,4 +1,4 @@
-package omock.model
+package omock.model.position
 
 @JvmInline
 value class Column(val comma: String) {
@@ -18,6 +18,6 @@ value class Column(val comma: String) {
         const val MIN_COLUMN_INDEX = 0
         const val MAX_COLUMN_INDEX = MAX_COLUMN.code - MIN_COLUMN.code
         val COLUM_RANGE = (MIN_COLUMN..MAX_COLUMN).map { it.toString() }
-        private const val ERROR_COLUMN_RANGE = "Column은 ${MIN_COLUMN}~${MAX_COLUMN} 사이어야 합니다."
+        private const val ERROR_COLUMN_RANGE = "Column은 $MIN_COLUMN~$MAX_COLUMN 사이어야 합니다."
     }
 }
