+++ src/main/kotlin/omock/model/search/DirectionResult.kt
@@ -1,4 +1,4 @@
-package omock.model
+package omock.model.search
 
 data class DirectionResult(
     val count: Int,
