+++ src/main/kotlin/omock/model/search/Node.kt
@@ -1,3 +1,3 @@
-package omock.model
+package omock.model.search
 
 data class Node(val x: Int, val y: Int)
