+++ src/main/kotlin/omock/model/stone/Stone.kt
@@ -1,4 +1,10 @@
-package omock.model
+package omock.model.stone
+
+import omock.model.player.BlackPlayer
+import omock.model.player.Player
+import omock.model.player.WhitePlayer
+import omock.model.position.Column
+import omock.model.position.Row
 
 data class Stone(
     val row: Row,
