+++ src/main/kotlin/omock/model/stonestate/Black.kt
@@ -1,0 +1,5 @@
+package omock.model.stonestate
+
+import omock.model.stone.Stone
+
+class Black(private val stone: Stone) : Placed(stone)
