+++ src/main/kotlin/omock/model/stonestate/Clear.kt
@@ -1,0 +1,6 @@
+package omock.model.stonestate
+
+import omock.model.stone.Stone
+import omock.model.stonestate.UnPlaced
+
+class Clear(private val stone: Stone) : UnPlaced(stone)
