+++ src/main/kotlin/omock/model/stonestate/StoneState.kt
@@ -1,4 +1,6 @@
-package omock.model
+package omock.model.stonestate
+
+import omock.model.player.Player
 
 interface StoneState {
     fun put(player: Player): StoneState
