+++ src/main/kotlin/omock/model/stonestate/White.kt
@@ -1,0 +1,5 @@
+package omock.model.stonestate
+
+import omock.model.stone.Stone
+
+class White(private val stone: Stone) : Placed(stone)
