+++ src/test/kotlin/omock/BoardTest.kt
@@ -1,11 +1,11 @@
 package omock
 
-import omock.model.Board
-import omock.model.Column
-import omock.model.Direction
-import omock.model.Row
-import omock.model.Stone
-import omock.model.WhitePlayer
+import omock.model.board.Board
+import omock.model.position.Column
+import omock.model.search.Direction
+import omock.model.position.Row
+import omock.model.stone.Stone
+import omock.model.player.WhitePlayer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
