+++ src/test/kotlin/omock/ColumnTest.kt
@@ -1,6 +1,6 @@
 package omock
 
-import omock.model.Column
+import omock.model.position.Column
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
