+++ src/test/kotlin/omock/StoneTest.kt
@@ -1,8 +1,8 @@
 package omock
 
-import omock.model.Column
-import omock.model.Row
-import omock.model.Stone
+import omock.model.position.Column
+import omock.model.position.Row
+import omock.model.stone.Stone
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
